added testDB

+++ retry/retry.go
@@ -5,6 +5,6 @@ import (
 )
 
 //ExpSleep sleeps for 50<<n milliseconds.
-func ExpSleep(n int) {
-	time.Sleep(time.Millisecond * 50 << uint(n))
+func ExpSleep(n uint) {
+	time.Sleep(time.Millisecond * 50 << n)
 }
